+++ src/types/index.ts
@@ -2,33 +2,27 @@
 
 // Base types and interfaces
 export * from "./items/types";
-export * from "./items/base";
-export * from "./items/workbench";
-export * from "./items/recipe";
+export * from "./base";
+export * from "./workbench";
+export * from "./recipe";
 export * from "./items/quickuse";
 export * from "./items/gear";
-export * from "./items/quest";
+export * from "./quest";
 export * from "./items/weapon";
 
 // Re-export commonly used types for convenience
-export type { BaseItem } from "./items/base";
+export type { BaseItem } from "./base";
 export type { Item } from "./items/item";
 export type {
 	Workbench,
 	WorkbenchTier,
 	WorkbenchUpgradeSummary,
 	WorkbenchUpgradeSummaryItem,
-} from "./items/workbench";
-export type { Recipe, RecipeRequirement } from "./items/recipe";
+} from "./workbench";
+export type { Recipe, RecipeRequirement } from "./recipe";
 export type { QuickUseData, QuickUseStat, QuickUseCharge } from "./items/quickuse";
 export type { GearData, GearStat } from "./items/gear";
-export type {
-	QuestId,
-	QuestObjective,
-	QuestObjectiveLink,
-	QuestReward,
-	Quest,
-} from "./items/quest";
+export type { QuestId, QuestObjective, QuestObjectiveLink, QuestReward, Quest } from "./quest";
 export type {
 	Weapon,
 	WeaponUpgrade,
