+++ src/types/quest.ts
